sroiv1/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
BareNecessities==0.2.8
bcrypt==4.2.1
blinker==1.8.2
cachelib==0.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.14.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
gast==0.5.4
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
itsdangerous==2.2.0
jax==0.4.28
jaxlib==0.4.28
Jinja2==3.1.4
jiter==0.6.1
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Mail==2.1.0
Mako==1.3.8
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.3.2
multidict==6.1.0
namex==0.0.8
numpy==1.26.4
openai==0.28.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pdfkit==1.0.0
pillow==10.3.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.1
pyodbc==5.2.0
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==26.0.3
requests==2.31.0
rich==13.7.1
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
SQLAlchemy==2.0.36
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.2.1
yarl==1.15.2
tenacity>=8.2.3
